@@ -1,5 +1,3 @@
-from typing import Union
-
 import pytest
 from twyn.similarity.algorithm import (
     AbstractSimilarityAlgorithm,
@@ -11,7 +9,7 @@ from twyn.similarity.exceptions import DistanceAlgorithmError, ThresholdError
 
 class TestAbstractSimilarityAlgorithm:
     class DifferentLettersSimilarityAlgorithm(AbstractSimilarityAlgorithm):
-        def _run_algorithm(self, first_sequence: str, second_sequence: str) -> Union[float, int]:
+        def _run_algorithm(self, first_sequence: str, second_sequence: str) -> float | int:
             first_letters = set(first_sequence)
             second_letters = set(second_sequence)
             return len(first_letters.symmetric_difference(second_letters))
@@ -54,7 +52,7 @@ class TestEditDistance:
 
 class TestExceptions:
     class ExceptionAlgorithm(AbstractSimilarityAlgorithm):
-        def _run_algorithm(self, first_sequence: str, second_sequence: str) -> Union[float, int]:
+        def _run_algorithm(self, first_sequence: str, second_sequence: str) -> float | int:
             raise KeyError
 
     def test_exception(self):
